pyproject.toml
[tool.poetry]
name = "sheets_and_friends"
version = "0.1.1"
description = "Create a LinkML model with as-is imported slots, imported but modified slots (via yq), or newly minted slots (via schemasheets)"
authors = ["Mark Andrew Miller <MAM@lbl.gov>"]
license = "MIT"

[tool.poetry.dependencies]
black = "*"
click_log = "*"
dpath = "^2.0.6"
glom = "^22.1.0"
hident = "^0.1.7"
pandasql = "^0.7.3"
pygsheets = "^2.0.5"
python = "^3.9"
schemasheets = "^0.1.8"
## using submodule while waiting for schemasheets examples fix to be merged and published
#linkml = "*"
deepdiff = "^5.7.0"
oaklib = "^0.1.22"
flake8 = "^4.0.1"
autoflake = "^1.4"
mypy = "^0.961"
typer = "^0.4.1"
strsimpy = "^0.2.1"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
do_shuttle = "sheets_and_friends.shuttle:do_shuttle"
modifications_and_validation = "sheets_and_friends.modifications_and_validation:modifications_and_validation"
#linkml2dataharmonizer = "sheets_and_friends.converters.commands:linkml2dataharmonizer"
#promote_to_select = "sheets_and_friends.converters.commands:promote_to_select"
compare_enums = "sheets_and_friends.compare_enums:compare_enums"
